Yang Yijie A3 - Anti-Fraud Unity2D Storybook
This report aims to summarize the key aspects of the anti-fraud Unity2D storytelling project, including the user interface, user experience, storyboarding techniques, prototyping techniques, and functionality utilizing the Unity2D engine. Considering that the Anti-Fraud Unity2D Storybook is an educational application based on the Unity2D engine, I provide an immersive 2D virtual storybook experience to educate users about fraud prevention. The theme of the storybook revolves around an online scam in which a young boy is victimized.The Unity2D Anti-Fraud Storybook was inspired by my concern for social issues, especially the impact of online scams and frauds on people.